minimal_requirements.txt
# Minimal requirements for Crypto Trading AI
# Install these manually if automated installation fails

# Core packages (try system packages first)
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Essential for the application
streamlit>=1.20.0
yfinance>=0.2.0
plotly>=5.0.0
requests>=2.25.0

# Technical analysis (try talib-binary if TA-Lib fails)
talib-binary>=0.4.0

# Machine learning (CPU version is fine)
tensorflow-cpu>=2.10.0

# Utilities
joblib>=1.0.0
tqdm>=4.60.0
